+++ lib/dp_exchange/coinbase/prime.ex
@@ -307,7 +307,14 @@ defmodule DpExchange.Coinbase.Prime do
 
   defp request_opts(opts) do
     opts
-    |> Keyword.take([:timeout, :retry_attempts, :retry_delay, :plug, :weight])
+    |> Keyword.take([
+      :timeout,
+      :retry_attempts,
+      :retry_delay,
+      :plug,
+      :weight,
+      :rate_limit_blocking
+    ])
     |> Keyword.put(:provider, :coinbase)
     |> Keyword.put_new(:limiter, Keyword.get(opts, :limiter, DpExchange.Coinbase.RateLimiter))
   end